refactor: add BoardController
