+++ src/components/task-edit.js
@@ -145,6 +145,6 @@ export default class TaskEdit extends AbstractComponent {
 
   setSubmitHandler(handler) {
     this.getElement().querySelector(`form`)
-      .addEventListener(`submit`, handler);
+    .addEventListener(`submit`, handler);
   }
 }
